gpu_app/requirements.txt
addict==2.4.0
affine==2.4.0
altair==6.0.0
aniso8601==10.0.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
anywidget==0.9.21
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
bqplot==0.12.45
branca==0.8.2
cachelib==0.13.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-plugins==1.1.1.2
cligj==0.7.2
cmap==0.7.0
color-operations==0.2.0
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
duckdb==1.4.3
eval_type_backport==0.3.1
exceptiongroup==1.3.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.2
fiona==1.10.1
Flask==3.1.2
Flask-Caching==2.3.1
flask-cors==6.0.2
flask-restx==1.3.2
flatbuffers==25.12.19
folium==0.20.0
fonttools==4.61.1
fqdn==1.5.1
fsspec==2025.12.0
gdown==5.2.0
geojson==3.2.0
geopandas==1.1.2
gitdb==4.0.12
GitPython==3.1.46
groundingdino-py==0.4.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.11
ImageIO==2.37.2
imgviz==1.8.0
importlib_resources==6.5.2
iopath==0.1.10
ipyevents==2.0.4
ipyfilechooser==0.6.0
ipykernel==7.1.0
ipyleaflet==0.20.0
ipympl==0.9.8
ipython==8.38.0
ipytree==0.2.2
ipyvue==1.11.3
ipyvuetify==1.11.3
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-leaflet==0.20.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
labelImg==1.8.6
labelme==5.10.1
lark==1.3.1
lazy_loader==0.4
leafmap==0.57.10
localtileserver==0.10.7
loguru==0.7.3
lxml==6.0.2
maplibre==0.3.6
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mistune==3.2.0
morecantile==6.2.0
mpmath==1.3.0
narwhals==2.15.0
natsort==8.4.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook_shim==0.2.4
numexpr==2.14.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
onnxruntime==1.23.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
osam==0.2.5
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
patool==4.0.1
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
plotly==6.5.1
portalocker==3.2.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==6.33.4
psutil==7.2.1
psygnal==0.15.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.0
pycocotools==2.0.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyogrio==0.12.1
pyparsing==3.3.1
pyproj==3.7.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.18.0
PySocks==1.7.1
pystac==1.14.2
pystac-client==0.9.0
python-box==7.3.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
rasterio==1.4.4
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rio-cogeo==6.0.0
rio-tiler==7.9.2
rioxarray==0.19.0
rpds-py==0.30.0
safetensors==0.7.0
sam2==1.1.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
scooby==0.11.0
segment-anything==1.0
segment-anything-hq==0.3
segment-anything-py==1.0.1
segment-geospatial==0.12.6
Send2Trash==2.0.0
server-thread==0.3.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
soupsieve==2.8.1
stack-data==0.6.3
streamlit==1.53.1
supervision==0.6.0
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.24
tinycss2==1.4.0
tokenizers==0.22.2
toml==0.10.2
tomli==2.3.0
torch==2.9.1
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
traittypes==0.2.3
transformers==4.57.3
triton==3.5.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
uvicorn==0.40.0
watchdog==6.0.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.5
whitebox==2.3.6
whiteboxgui==2.3.0
widgetsnbextension==4.0.15
xarray==2025.6.1
xyzservices==2025.11.0
yapf==0.43.0
